+++ hrms/hr/doctype/employee_checkin/employee_checkin.js
@@ -16,65 +16,7 @@ frappe.ui.form.on("Employee Checkin", {
 		}
 	},
 
-	add_fetch_shift_button(frm) {
-		if (frm.doc.attendace) return;
-		frm.add_custom_button(__("Fetch Shift"), function () {
-			const previous_shift = frm.doc.shift;
-			frappe.call({
-				method: "fetch_shift",
-				doc: frm.doc,
-				freeze: true,
-				freeze_message: __("Fetching Shift"),
-				callback: function () {
-					if (previous_shift === frm.doc.shift) return;
-					frm.dirty();
-					frm.save();
-					frappe.show_alert({
-						message: __("Shift has been successfully updated to {0}.", [
-							frm.doc.shift,
-						]),
-						indicator: "green",
-					});
-				},
-			});
-		});
-	},
-
-	fetch_geolocation: async (frm) => {
-		if (!navigator.geolocation) {
-			frappe.msgprint({
-				message: __("Geolocation is not supported by your current browser"),
-				title: __("Geolocation Error"),
-				indicator: "red",
-			});
-			hide_field(["geolocation"]);
-			return;
-		}
-
-		frappe.dom.freeze(__("Fetching your geolocation") + "...");
-
-		navigator.geolocation.getCurrentPosition(
-			async (position) => {
-				frm.set_value("latitude", position.coords.latitude);
-				frm.set_value("longitude", position.coords.longitude);
-
-				await frm.call("set_geolocation_from_coordinates");
-				frm.dirty();
-				frappe.dom.unfreeze();
-			},
-			(error) => {
-				frappe.dom.unfreeze();
-
-				let msg = __("Unable to retrieve your location") + "<br><br>";
-				if (error) {
-					msg += __("ERROR({0}): {1}", [error.code, error.message]);
-				}
-				frappe.msgprint({
-					message: msg,
-					title: __("Geolocation Error"),
-					indicator: "red",
-				});
-			},
-		);
+	fetch_geolocation: (frm) => {
+		hrms.fetch_geolocation(frm);
 	},
 });
