+++ hrms/hr/doctype/shift_location/shift_location.js
@@ -1,8 +1,26 @@
 // Copyright (c) 2024, Frappe Technologies Pvt. Ltd. and contributors
 // For license information, please see license.txt
 
-// frappe.ui.form.on("Shift Location", {
-// 	refresh(frm) {
+frappe.ui.form.on("Shift Location", {
+	refresh: async () => {
+		const allow_geolocation_tracking = await frappe.db.get_single_value(
+			"HR Settings",
+			"allow_geolocation_tracking",
+		);
 
-// 	},
-// });
+		if (!allow_geolocation_tracking) {
+			hide_field([
+				"checkin_radius",
+				"fetch_geolocation",
+				"latitude",
+				"longitude",
+				"geolocation",
+			]);
+			return;
+		}
+	},
+
+	fetch_geolocation: (frm) => {
+		hrms.fetch_geolocation(frm);
+	},
+});
