+++ hrms/public/js/utils/index.js
@@ -179,20 +179,40 @@ $.extend(hrms, {
 		});
 	},
 
-	get_doctype_fields_for_autocompletion: (doctype) => {
-		const fields = frappe.get_meta(doctype).fields;
-		const autocompletions = [];
-
-		fields
-			.filter((df) => !frappe.model.no_value_type.includes(df.fieldtype))
-			.map((df) => {
-				autocompletions.push({
-					value: df.fieldname,
-					score: 8,
-					meta: __("{0} Field", [doctype]),
-				});
+	fetch_geolocation: async (frm) => {
+		if (!navigator.geolocation) {
+			frappe.msgprint({
+				message: __("Geolocation is not supported by your current browser"),
+				title: __("Geolocation Error"),
+				indicator: "red",
 			});
+			hide_field(["geolocation"]);
+			return;
+		}
 
-		return autocompletions;
+		frappe.dom.freeze(__("Fetching your geolocation") + "...");
+
+		navigator.geolocation.getCurrentPosition(
+			async (position) => {
+				frm.set_value("latitude", position.coords.latitude);
+				frm.set_value("longitude", position.coords.longitude);
+
+				await frm.call("hrms.hr.utils.set_geolocation_from_coordinates");
+				frappe.dom.unfreeze();
+			},
+			(error) => {
+				frappe.dom.unfreeze();
+
+				let msg = __("Unable to retrieve your location") + "<br><br>";
+				if (error) {
+					msg += __("ERROR({0}): {1}", [error.code, error.message]);
+				}
+				frappe.msgprint({
+					message: msg,
+					title: __("Geolocation Error"),
+					indicator: "red",
+				});
+			},
+		);
 	},
 });
